@@ -105,7 +105,7 @@
                             <th>Administrator</th>
                             <th>Device Groups</th>
                             <th>Reset Password</th>
-                            <th>Delete Group</th>
+                            <th>Delete User</th>
                         </tr>
                     </thead>
                     <tbody>
