@@ -2,9 +2,9 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Primitives;
 using Remotely.Server.Data;
 using Remotely.Server.Models;
 using Remotely.Shared.Enums;
@@ -13,7 +13,6 @@ using Remotely.Shared.Utilities;
 using Remotely.Shared.ViewModels;
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -23,6 +22,7 @@ namespace Remotely.Server.Services
     public interface IDataService
     {
         Task AddAlert(string deviceID, string organizationID, string alertMessage, string details = null);
+
         bool AddDeviceGroup(string orgID, DeviceGroup deviceGroup, out string deviceGroupID, out string errorMessage);
 
         InviteLink AddInvite(string orgID, InviteViewModel invite);
@@ -34,6 +34,8 @@ namespace Remotely.Server.Services
         void AddOrUpdateScriptResult(ScriptResult scriptResult);
 
         Task AddOrUpdateScriptSchedule(ScriptSchedule schedule);
+
+        Task AddScriptResultToScriptRun(string scriptResultId, int scriptRunId);
 
         Task AddScriptRun(ScriptRun scriptRun);
 
@@ -48,6 +50,7 @@ namespace Remotely.Server.Services
         void CleanupOldRecords();
 
         Task ClearLogs(string currentUserName);
+
         Task<ApiToken> CreateApiToken(string userName, string tokenName, string secretHash);
 
         Task<Device> CreateDevice(DeviceSetupOptions options);
@@ -79,9 +82,9 @@ namespace Remotely.Server.Services
         bool DoesUserHaveAccessToDevice(string deviceID, RemotelyUser remotelyUser);
 
         bool DoesUserHaveAccessToDevice(string deviceID, string remotelyUserID);
-        Task<DeviceGroup> GetDeviceGroup(string deviceGroupID);
+
         string[] FilterDeviceIDsByUserPermission(string[] deviceIDs, RemotelyUser remotelyUser);
-        Task AddScriptResultToScriptRun(string scriptResultId, int scriptRunId);
+
         string[] FilterUsersByDevicePermission(IEnumerable<string> userIDs, string deviceID);
 
         Task<Alert> GetAlert(string alertID);
@@ -124,6 +127,7 @@ namespace Remotely.Server.Services
 
         int GetDeviceCount(RemotelyUser user);
 
+        Task<DeviceGroup> GetDeviceGroup(string deviceGroupID);
         DeviceGroup[] GetDeviceGroups(string username);
 
         DeviceGroup[] GetDeviceGroupsForOrganization(string organizationId);
@@ -163,6 +167,7 @@ namespace Remotely.Server.Services
         Task<List<ScriptSchedule>> GetScriptSchedules(string organizationID);
 
         Task<List<ScriptSchedule>> GetScriptSchedulesDue();
+
         List<string> GetServerAdmins();
 
         SharedFile GetSharedFiled(string fileID);
@@ -184,6 +189,7 @@ namespace Remotely.Server.Services
         void RemoveDevices(string[] deviceIDs);
 
         Task<bool> RemoveUserFromDeviceGroup(string orgID, string groupID, string userID);
+
         Task RenameApiToken(string userName, string tokenId, string tokenName);
 
         Task ResetBranding(string organizationId);
@@ -202,43 +208,54 @@ namespace Remotely.Server.Services
 
         Task UpdateBrandingInfo(
                                                                                                                                                                                                                                                     string organizationId,
-            string productName, 
+            string productName,
             byte[] iconBytes,
-            ColorPickerModel titleForeground, 
-            ColorPickerModel titleBackground, 
+            ColorPickerModel titleForeground,
+            ColorPickerModel titleBackground,
             ColorPickerModel titleButtonForeground);
+
         Task<Device> UpdateDevice(DeviceSetupOptions deviceOptions, string organizationId);
+
         void UpdateDevice(string deviceID, string tag, string alias, string deviceGroupID, string notes, WebRtcSetting webRtcSetting);
+
         void UpdateOrganizationName(string orgID, string organizationName);
+
         void UpdateTags(string deviceID, string tags);
+
         void UpdateUserOptions(string userName, RemotelyUserOptions options);
+
         bool ValidateApiKey(string keyId, string apiSecret, string requestPath, string remoteIP);
+
         void WriteEvent(EventLog eventLog);
+
         void WriteEvent(Exception ex, string organizationID);
+
         void WriteEvent(string message, EventType eventType, string organizationID);
+
         void WriteEvent(string message, string organizationID);
+
         void WriteLog(LogLevel logLevel, string category, EventId eventId, string state, Exception exception, string[] scopeStack);
     }
 
     public class DataService : IDataService
     {
         private readonly IApplicationConfig _appConfig;
-
-        private readonly IDbContextFactory<AppDb> _dbFactory;
         private readonly IHostEnvironment _hostEnvironment;
+        private readonly IAppDbFactory _appDbFactory;
 
-        public DataService(IDbContextFactory<AppDb> dbFactory,
+        public DataService(
             IApplicationConfig appConfig,
-            IHostEnvironment hostEnvironment)
+            IHostEnvironment hostEnvironment,
+            IAppDbFactory appDbFactory)
         {
-            _dbFactory = dbFactory;
             _appConfig = appConfig;
             _hostEnvironment = hostEnvironment;
+            _appDbFactory = appDbFactory;
         }
 
         public async Task AddAlert(string deviceId, string organizationID, string alertMessage, string details = null)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var users = dbContext.Users
                .Include(x => x.Alerts)
@@ -272,7 +289,7 @@ namespace Remotely.Server.Services
 
         public bool AddDeviceGroup(string orgID, DeviceGroup deviceGroup, out string deviceGroupID, out string errorMessage)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             deviceGroupID = null;
             errorMessage = null;
@@ -301,7 +318,7 @@ namespace Remotely.Server.Services
 
         public InviteLink AddInvite(string orgID, InviteViewModel invite)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var organization = dbContext.Organizations
                 .Include(x => x.InviteLinks)
@@ -323,7 +340,7 @@ namespace Remotely.Server.Services
 
         public bool AddOrUpdateDevice(Device device, out Device updatedDevice)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var existingDevice = dbContext.Devices.Find(device.ID);
             if (existingDevice != null)
@@ -378,7 +395,7 @@ namespace Remotely.Server.Services
 
         public async Task AddOrUpdateSavedScript(SavedScript script, string userId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             dbContext.SavedScripts.Update(script);
             script.CreatorId = userId;
@@ -389,7 +406,7 @@ namespace Remotely.Server.Services
 
         public void AddOrUpdateScriptResult(ScriptResult result)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var device = dbContext.Devices.Find(result.DeviceID);
 
@@ -416,7 +433,7 @@ namespace Remotely.Server.Services
 
         public async Task AddOrUpdateScriptSchedule(ScriptSchedule schedule)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var existingSchedule = await dbContext.ScriptSchedules
                 .Include(x => x.Creator)
@@ -459,7 +476,7 @@ namespace Remotely.Server.Services
 
         public async Task AddScriptResultToScriptRun(string scriptResultId, int scriptRunId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var run = await dbContext.ScriptRuns
                 .Include(x => x.Results)
@@ -488,7 +505,7 @@ namespace Remotely.Server.Services
 
         public async Task AddScriptRun(ScriptRun scriptRun)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             dbContext.Attach(scriptRun);
             dbContext.ScriptRuns.Add(scriptRun);
@@ -522,7 +539,7 @@ namespace Remotely.Server.Services
 
         public bool AddUserToDeviceGroup(string orgID, string groupID, string userName, out string resultMessage)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             resultMessage = string.Empty;
 
@@ -570,7 +587,7 @@ namespace Remotely.Server.Services
 
         public void ChangeUserIsAdmin(string organizationID, string targetUserID, bool isAdmin)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var targetUser = dbContext.Users.FirstOrDefault(x =>
                                 x.OrganizationID == organizationID &&
@@ -585,17 +602,16 @@ namespace Remotely.Server.Services
 
         public void CleanupOldRecords()
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             if (_appConfig.DataRetentionInDays > -1)
             {
-
                 var expirationDate = DateTimeOffset.Now - TimeSpan.FromDays(_appConfig.DataRetentionInDays);
 
                 var scriptRuns = dbContext.ScriptRuns
-                    .Include(x=>x.Results)
-                    .Include(x=>x.Devices)
-                    .Include(x=>x.DevicesCompleted)
+                    .Include(x => x.Results)
+                    .Include(x => x.Devices)
+                    .Include(x => x.DevicesCompleted)
                     .Where(x => x.RunAt < expirationDate);
 
                 foreach (var run in scriptRuns)
@@ -628,7 +644,7 @@ namespace Remotely.Server.Services
 
         public async Task ClearLogs(string currentUserName)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var currentUser = await dbContext.Users.FirstOrDefaultAsync(x => x.UserName == currentUserName);
             if (currentUser is null)
@@ -638,7 +654,6 @@ namespace Remotely.Server.Services
 
             try
             {
-
                 if (currentUser.IsServerAdmin)
                 {
                     dbContext.EventLogs.RemoveRange(dbContext.EventLogs);
@@ -659,7 +674,7 @@ namespace Remotely.Server.Services
 
         public async Task<ApiToken> CreateApiToken(string userName, string tokenName, string secretHash)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var user = dbContext.Users.FirstOrDefault(x => x.UserName == userName);
 
@@ -676,7 +691,7 @@ namespace Remotely.Server.Services
 
         public async Task<Device> CreateDevice(DeviceSetupOptions options)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             try
             {
@@ -718,12 +733,11 @@ namespace Remotely.Server.Services
                 WriteEvent(ex, options.OrganizationID);
                 return null;
             }
-
         }
 
         public async Task<bool> CreateUser(string userEmail, bool isAdmin, string organizationID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             try
             {
@@ -748,12 +762,11 @@ namespace Remotely.Server.Services
                 WriteEvent(ex, organizationID);
                 return false;
             }
-
         }
 
         public async Task DeleteAlert(Alert alert)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             dbContext.Alerts.Remove(alert);
             await dbContext.SaveChangesAsync();
@@ -761,7 +774,7 @@ namespace Remotely.Server.Services
 
         public async Task DeleteAllAlerts(string orgID, string userName = null)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var alerts = dbContext.Alerts.Where(x => x.OrganizationID == orgID);
 
@@ -778,7 +791,7 @@ namespace Remotely.Server.Services
 
         public async Task DeleteApiToken(string userName, string tokenId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var user = dbContext.Users.FirstOrDefault(x => x.UserName == userName);
             var token = dbContext.ApiTokens.FirstOrDefault(x =>
@@ -791,7 +804,7 @@ namespace Remotely.Server.Services
 
         public void DeleteDeviceGroup(string orgID, string deviceGroupID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var deviceGroup = dbContext.DeviceGroups
                 .Include(x => x.Devices)
@@ -821,7 +834,7 @@ namespace Remotely.Server.Services
 
         public void DeleteInvite(string orgID, string inviteID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var invite = dbContext.InviteLinks.FirstOrDefault(x =>
                 x.OrganizationID == orgID &&
@@ -839,7 +852,7 @@ namespace Remotely.Server.Services
 
         public async Task DeleteSavedScript(Guid scriptId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var script = dbContext.SavedScripts.Find(scriptId);
             if (script is not null)
@@ -851,7 +864,7 @@ namespace Remotely.Server.Services
 
         public async Task DeleteScriptSchedule(int scriptScheduleId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var schedule = dbContext.ScriptSchedules.Find(scriptScheduleId);
             if (schedule is not null)
@@ -863,7 +876,7 @@ namespace Remotely.Server.Services
 
         public async Task DeleteUser(string orgID, string targetUserID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var target = dbContext.Users
                 .Include(x => x.DeviceGroups)
@@ -903,24 +916,21 @@ namespace Remotely.Server.Services
                 .FirstOrDefault(x => x.ID == orgID)
                 .RemotelyUsers.Remove(target);
 
-
             dbContext.Users.Remove(target);
 
-
             await dbContext.SaveChangesAsync();
-
         }
 
         public void DetachEntity(object entity)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             dbContext.Entry(entity).State = EntityState.Detached;
         }
 
         public void DeviceDisconnected(string deviceID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var device = dbContext.Devices.Find(deviceID);
             if (device != null)
@@ -933,7 +943,7 @@ namespace Remotely.Server.Services
 
         public bool DoesUserExist(string userName)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             if (string.IsNullOrWhiteSpace(userName))
             {
@@ -949,7 +959,7 @@ namespace Remotely.Server.Services
                 return false;
             }
 
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Devices
                 .Include(x => x.DeviceGroup)
@@ -965,7 +975,7 @@ namespace Remotely.Server.Services
 
         public bool DoesUserHaveAccessToDevice(string deviceID, string remotelyUserID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var remotelyUser = dbContext.Users.Find(remotelyUserID);
 
@@ -974,7 +984,7 @@ namespace Remotely.Server.Services
 
         public string[] FilterDeviceIDsByUserPermission(string[] deviceIDs, RemotelyUser remotelyUser)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Devices
                 .Include(x => x.DeviceGroup)
@@ -993,14 +1003,14 @@ namespace Remotely.Server.Services
 
         public string[] FilterUsersByDevicePermission(IEnumerable<string> userIDs, string deviceID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return FilterUsersByDevicePermissionInternal(dbContext, userIDs, deviceID);
         }
 
         public async Task<Alert> GetAlert(string alertID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return await dbContext.Alerts
                 .Include(x => x.Device)
@@ -1010,7 +1020,7 @@ namespace Remotely.Server.Services
 
         public Alert[] GetAlerts(string userID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Alerts
                 .Include(x => x.Device)
@@ -1022,7 +1032,7 @@ namespace Remotely.Server.Services
 
         public ApiToken[] GetAllApiTokens(string userID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var user = dbContext.Users.FirstOrDefault(x => x.Id == userID);
 
@@ -1034,7 +1044,7 @@ namespace Remotely.Server.Services
 
         public ScriptResult[] GetAllCommandResults(string orgID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.ScriptResults
                 .Where(x => x.OrganizationID == orgID)
@@ -1044,7 +1054,7 @@ namespace Remotely.Server.Services
 
         public ScriptResult[] GetAllCommandResultsForUser(string orgId, string userName, string deviceId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.ScriptResults
                 .Where(x => x.OrganizationID == orgId &&
@@ -1056,14 +1066,14 @@ namespace Remotely.Server.Services
 
         public Device[] GetAllDevices(string orgID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Devices.Where(x => x.OrganizationID == orgID).ToArray();
         }
 
         public EventLog[] GetAllEventLogs(string orgID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.EventLogs
                 .Where(x => x.OrganizationID == orgID)
@@ -1073,7 +1083,7 @@ namespace Remotely.Server.Services
 
         public InviteLink[] GetAllInviteLinks(string organizationId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.InviteLinks
                 .Where(x => x.OrganizationID == organizationId)
@@ -1082,7 +1092,7 @@ namespace Remotely.Server.Services
 
         public ScriptResult[] GetAllScriptResults(string orgId, string deviceId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.ScriptResults
                 .Where(x => x.OrganizationID == orgId && x.DeviceID == deviceId)
@@ -1092,7 +1102,7 @@ namespace Remotely.Server.Services
 
         public ScriptResult[] GetAllScriptResultsForUser(string orgId, string userName)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.ScriptResults
                 .Where(x => x.OrganizationID == orgId && x.SenderUserName == userName)
@@ -1102,7 +1112,7 @@ namespace Remotely.Server.Services
 
         public RemotelyUser[] GetAllUsersForServer()
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Users.ToArray();
         }
@@ -1114,7 +1124,7 @@ namespace Remotely.Server.Services
                 return Array.Empty<RemotelyUser>();
             }
 
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var organization = await dbContext.Organizations
                 .Include(x => x.RemotelyUsers)
@@ -1130,7 +1140,7 @@ namespace Remotely.Server.Services
                 return null;
             }
 
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.ApiTokens.FirstOrDefault(x => x.ID == keyId);
         }
@@ -1142,7 +1152,7 @@ namespace Remotely.Server.Services
                 return null;
             }
 
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var organization = await dbContext.Organizations
               .Include(x => x.BrandingInfo)
@@ -1163,14 +1173,14 @@ namespace Remotely.Server.Services
 
         public async Task<Organization> GetDefaultOrganization()
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return await dbContext.Organizations.FirstOrDefaultAsync(x => x.IsDefaultOrganization);
         }
 
         public async Task<string> GetDefaultRelayCode()
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var relayCode = await dbContext.Organizations
                 .Where(x => x.IsDefaultOrganization)
@@ -1182,7 +1192,7 @@ namespace Remotely.Server.Services
 
         public Device GetDevice(string orgID, string deviceID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Devices.FirstOrDefault(x =>
                             x.OrganizationID == orgID &&
@@ -1191,21 +1201,21 @@ namespace Remotely.Server.Services
 
         public Device GetDevice(string deviceID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Devices.FirstOrDefault(x => x.ID == deviceID);
         }
 
         public int GetDeviceCount()
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Devices.Count();
         }
 
         public int GetDeviceCount(RemotelyUser user)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Devices
                 .Include(x => x.DeviceGroup)
@@ -1222,13 +1232,13 @@ namespace Remotely.Server.Services
 
         public async Task<DeviceGroup> GetDeviceGroup(string deviceGroupID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
             return await dbContext.DeviceGroups.FindAsync(deviceGroupID);
         }
 
         public DeviceGroup[] GetDeviceGroups(string username)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var user = dbContext.Users.FirstOrDefault(x => x.UserName == username);
 
@@ -1264,7 +1274,7 @@ namespace Remotely.Server.Services
 
         public DeviceGroup[] GetDeviceGroupsForOrganization(string organizationId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.DeviceGroups
                 .Include(x => x.Users)
@@ -1276,17 +1286,16 @@ namespace Remotely.Server.Services
 
         public List<Device> GetDevices(IEnumerable<string> deviceIds)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Devices
                 .Where(x => deviceIds.Contains(x.ID))
                 .ToList();
-
         }
 
         public Device[] GetDevicesForUser(string userName)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             if (string.IsNullOrWhiteSpace(userName))
             {
@@ -1320,7 +1329,7 @@ namespace Remotely.Server.Services
 
         public EventLog[] GetEventLogs(string userName, DateTimeOffset from, DateTimeOffset to, EventType? type, string message)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var user = dbContext.Users
                         .FirstOrDefault(x => x.UserName == userName);
@@ -1341,7 +1350,7 @@ namespace Remotely.Server.Services
             {
                 var orgID = user.OrganizationID;
                 query = query
-                        .Where(x => x.OrganizationID == orgID && 
+                        .Where(x => x.OrganizationID == orgID &&
                             x.TimeStamp >= fromDate && x.TimeStamp <= toDate)
                         .OrderByDescending(x => x.TimeStamp);
             }
@@ -1359,14 +1368,14 @@ namespace Remotely.Server.Services
 
         public Organization GetOrganizationById(string organizationID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Organizations.Find(organizationID);
         }
 
         public async Task<Organization> GetOrganizationByRelayCode(string relayCode)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             if (string.IsNullOrWhiteSpace(relayCode))
             {
@@ -1378,7 +1387,7 @@ namespace Remotely.Server.Services
 
         public async Task<Organization> GetOrganizationByUserName(string userName)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var user = await dbContext
                 .Users
@@ -1390,21 +1399,21 @@ namespace Remotely.Server.Services
 
         public int GetOrganizationCount()
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Organizations.Count();
         }
 
         public string GetOrganizationNameById(string organizationID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Organizations.FirstOrDefault(x => x.ID == organizationID)?.OrganizationName;
         }
 
         public string GetOrganizationNameByUserName(string userName)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Users
                    .Include(x => x.Organization)
@@ -1415,12 +1424,12 @@ namespace Remotely.Server.Services
 
         public async Task<List<ScriptRun>> GetPendingScriptRuns(string deviceId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var pendingRuns = new List<ScriptRun>();
 
             var now = Time.Now;
-    
+
             var scriptRunGroups = dbContext.ScriptRuns
                 .Include(x => x.Devices)
                 .Include(x => x.DevicesCompleted)
@@ -1446,13 +1455,12 @@ namespace Remotely.Server.Services
 
             await dbContext.SaveChangesAsync();
 
-
             return pendingRuns;
         }
 
         public async Task<List<SavedScript>> GetQuickScripts(string userId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return await dbContext.SavedScripts
                 .Where(x => x.CreatorId == userId && x.IsQuickScript)
@@ -1461,8 +1469,8 @@ namespace Remotely.Server.Services
 
         public async Task<SavedScript> GetSavedScript(string userId, Guid scriptId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
-            
+            using var dbContext = _appDbFactory.GetContext();
+
             return await dbContext.SavedScripts
                 .FirstOrDefaultAsync(x =>
                     x.Id == scriptId &&
@@ -1471,13 +1479,13 @@ namespace Remotely.Server.Services
 
         public async Task<SavedScript> GetSavedScript(Guid scriptId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
             return await dbContext.SavedScripts.FirstOrDefaultAsync(x => x.Id == scriptId);
         }
 
         public async Task<List<SavedScript>> GetSavedScriptsWithoutContent(string userId, string organizationId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var query = dbContext.SavedScripts
                 .Include(x => x.Creator)
@@ -1499,7 +1507,7 @@ namespace Remotely.Server.Services
 
         public ScriptResult GetScriptResult(string commandResultID, string orgID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.ScriptResults
                 .FirstOrDefault(x =>
@@ -1509,14 +1517,14 @@ namespace Remotely.Server.Services
 
         public ScriptResult GetScriptResult(string commandResultID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.ScriptResults.Find(commandResultID);
         }
 
         public async Task<List<ScriptSchedule>> GetScriptSchedules(string organizationId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
             return await dbContext.ScriptSchedules
                 .Include(x => x.Creator)
                 .Include(x => x.Devices)
@@ -1527,7 +1535,7 @@ namespace Remotely.Server.Services
 
         public async Task<List<ScriptSchedule>> GetScriptSchedulesDue()
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var now = Time.Now;
 
@@ -1541,7 +1549,7 @@ namespace Remotely.Server.Services
 
         public List<string> GetServerAdmins()
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Users
                 .Where(x => x.IsServerAdmin)
@@ -1551,14 +1559,14 @@ namespace Remotely.Server.Services
 
         public SharedFile GetSharedFiled(string fileID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.SharedFiles.Find(fileID);
         }
 
         public int GetTotalDevices()
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Devices.Count();
         }
@@ -1569,7 +1577,7 @@ namespace Remotely.Server.Services
             {
                 return null;
             }
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return await dbContext.Users.FirstOrDefaultAsync(x => x.UserName == username);
         }
@@ -1580,7 +1588,7 @@ namespace Remotely.Server.Services
             {
                 return null;
             }
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Users.FirstOrDefault(x => x.Id == userID);
         }
@@ -1592,7 +1600,7 @@ namespace Remotely.Server.Services
                 return null;
             }
 
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Users
                 .Include(x => x.Organization)
@@ -1601,7 +1609,7 @@ namespace Remotely.Server.Services
 
         public RemotelyUserOptions GetUserOptions(string userName)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             return dbContext.Users
                     .FirstOrDefault(x => x.UserName == userName)
@@ -1610,7 +1618,7 @@ namespace Remotely.Server.Services
 
         public bool JoinViaInvitation(string userName, string inviteID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var invite = dbContext.InviteLinks.FirstOrDefault(x =>
                             x.InvitedUser.ToLower() == userName.ToLower() &&
@@ -1640,7 +1648,7 @@ namespace Remotely.Server.Services
 
         public void PopulateRelayCodes()
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             foreach (var organization in dbContext.Organizations)
             {
@@ -1658,7 +1666,7 @@ namespace Remotely.Server.Services
 
         public void RemoveDevices(string[] deviceIDs)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var devices = dbContext.Devices
                 .Include(x => x.ScriptResults)
@@ -1675,7 +1683,7 @@ namespace Remotely.Server.Services
 
         public async Task<bool> RemoveUserFromDeviceGroup(string orgID, string groupID, string userID)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var deviceGroup = dbContext.DeviceGroups
                 .Include(x => x.Users)
@@ -1699,7 +1707,7 @@ namespace Remotely.Server.Services
 
         public async Task RenameApiToken(string userName, string tokenId, string tokenName)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var user = dbContext.Users.FirstOrDefault(x => x.UserName == userName);
             var token = dbContext.ApiTokens.FirstOrDefault(x =>
@@ -1712,7 +1720,7 @@ namespace Remotely.Server.Services
 
         public async Task ResetBranding(string organizationId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var organization = await dbContext.Organizations
                .Include(x => x.BrandingInfo)
@@ -1730,7 +1738,7 @@ namespace Remotely.Server.Services
 
         public void SetAllDevicesNotOnline()
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             dbContext.Devices.ForEachAsync(x =>
             {
@@ -1741,7 +1749,7 @@ namespace Remotely.Server.Services
 
         public async Task SetDisplayName(RemotelyUser user, string displayName)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             dbContext.Attach(user);
             user.UserOptions.DisplayName = displayName;
@@ -1750,7 +1758,7 @@ namespace Remotely.Server.Services
 
         public async Task SetIsDefaultOrganization(string orgID, bool isDefault)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var organization = await dbContext.Organizations.FindAsync(orgID);
             if (organization is null)
@@ -1769,7 +1777,7 @@ namespace Remotely.Server.Services
 
         public async Task SetIsServerAdmin(string targetUserId, bool isServerAdmin, string callerUserId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var caller = await dbContext.Users.FindAsync(callerUserId);
             if (caller?.IsServerAdmin != true)
@@ -1796,7 +1804,7 @@ namespace Remotely.Server.Services
 
         public void SetServerVerificationToken(string deviceID, string verificationToken)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var device = dbContext.Devices.Find(deviceID);
             if (device != null)
@@ -1819,7 +1827,7 @@ namespace Remotely.Server.Services
             {
                 return false;
             }
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             if (user.TempPassword == password)
             {
@@ -1839,7 +1847,7 @@ namespace Remotely.Server.Services
             ColorPickerModel titleBackground,
             ColorPickerModel titleButtonForeground)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var organization = await dbContext.Organizations
                 .Include(x => x.BrandingInfo)
@@ -1879,16 +1887,28 @@ namespace Remotely.Server.Services
 
         public void UpdateDevice(string deviceID, string tag, string alias, string deviceGroupID, string notes, WebRtcSetting webRtcSetting)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
-            var device = dbContext.Devices.Find(deviceID);
+            var device = dbContext.Devices
+                .Include(x => x.DeviceGroup)
+                .FirstOrDefault(x => x.ID == deviceID);
             if (device == null)
             {
                 return;
             }
 
+            if (string.IsNullOrWhiteSpace(deviceGroupID))
+            {
+                device.DeviceGroup?.Devices?.RemoveAll(x => x.ID == deviceID);
+                device.DeviceGroup = null;
+                device.DeviceGroupID = null;
+            }
+            else
+            {
+                device.DeviceGroupID = deviceGroupID;
+            }
+
             device.Tags = tag;
-            device.DeviceGroupID = deviceGroupID;
             device.Alias = alias;
             device.Notes = notes;
             device.WebRtcSetting = webRtcSetting;
@@ -1897,7 +1917,7 @@ namespace Remotely.Server.Services
 
         public async Task<Device> UpdateDevice(DeviceSetupOptions deviceOptions, string organizationId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var device = dbContext.Devices.Find(deviceOptions.DeviceID);
             if (device == null || device.OrganizationID != organizationId)
@@ -1917,7 +1937,7 @@ namespace Remotely.Server.Services
 
         public void UpdateOrganizationName(string orgID, string organizationName)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             dbContext.Organizations
                 .FirstOrDefault(x => x.ID == orgID)
@@ -1927,7 +1947,7 @@ namespace Remotely.Server.Services
 
         public void UpdateTags(string deviceID, string tags)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var device = dbContext.Devices.Find(deviceID);
             if (device == null)
@@ -1941,7 +1961,7 @@ namespace Remotely.Server.Services
 
         public void UpdateUserOptions(string userName, RemotelyUserOptions options)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             dbContext.Users.FirstOrDefault(x => x.UserName == userName).UserOptions = options;
             dbContext.SaveChanges();
@@ -1949,15 +1969,13 @@ namespace Remotely.Server.Services
 
         public bool ValidateApiKey(string keyId, string apiSecret, string requestPath, string remoteIP)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var hasher = new PasswordHasher<RemotelyUser>();
             var token = dbContext.ApiTokens.FirstOrDefault(x => x.ID == keyId);
 
-
             var isValid = token is not null &&
                 hasher.VerifyHashedPassword(null, token.Secret, apiSecret) == PasswordVerificationResult.Success;
-
 
             if (token is not null)
             {
@@ -1974,7 +1992,7 @@ namespace Remotely.Server.Services
         {
             try
             {
-                using var dbContext = _dbFactory.CreateDbContext();
+                using var dbContext = _appDbFactory.GetContext();
 
                 dbContext.EventLogs.Add(eventLog);
                 dbContext.SaveChanges();
@@ -1986,7 +2004,7 @@ namespace Remotely.Server.Services
         {
             try
             {
-                using var dbContext = _dbFactory.CreateDbContext();
+                using var dbContext = _appDbFactory.GetContext();
 
                 dbContext.EventLogs.Add(new EventLog()
                 {
@@ -2011,7 +2029,7 @@ namespace Remotely.Server.Services
         {
             try
             {
-                using var dbContext = _dbFactory.CreateDbContext();
+                using var dbContext = _appDbFactory.GetContext();
 
                 dbContext.EventLogs.Add(new EventLog()
                 {
@@ -2036,7 +2054,7 @@ namespace Remotely.Server.Services
             try
             {
                 // TODO: Refactor EventLog to resemble these params.  Replace WriteEvent with ILogger<T>.
-                using var dbContext = _dbFactory.CreateDbContext();
+                using var dbContext = _appDbFactory.GetContext();
 
                 EventType eventType = EventType.Debug;
                 switch (logLevel)
@@ -2046,16 +2064,20 @@ namespace Remotely.Server.Services
                     case LogLevel.Debug:
                         eventType = EventType.Debug;
                         break;
+
                     case LogLevel.Information:
                         eventType = EventType.Info;
                         break;
+
                     case LogLevel.Warning:
                         eventType = EventType.Warning;
                         break;
+
                     case LogLevel.Error:
                     case LogLevel.Critical:
                         eventType = EventType.Error;
                         break;
+
                     default:
                         break;
                 }
@@ -2071,7 +2093,6 @@ namespace Remotely.Server.Services
                 dbContext.SaveChanges();
             }
             catch { }
-
         }
 
         private async Task<string> AddSharedFileInternal(
@@ -2080,7 +2101,7 @@ namespace Remotely.Server.Services
             string contentType,
             string organizationId)
         {
-            using var dbContext = _dbFactory.CreateDbContext();
+            using var dbContext = _appDbFactory.GetContext();
 
             var expirationDate = DateTimeOffset.Now.AddDays(-_appConfig.DataRetentionInDays);
             var expiredFiles = dbContext.SharedFiles.Where(x => x.Timestamp < expirationDate);
@@ -2099,6 +2120,7 @@ namespace Remotely.Server.Services
             await dbContext.SaveChangesAsync();
             return sharedFile.ID;
         }
+
         private string[] FilterUsersByDevicePermissionInternal(AppDb dbContext, IEnumerable<string> userIDs, string deviceID)
         {
             var device = dbContext.Devices
