@@ -158,8 +158,11 @@ export class MessageSender {
         if (ViewerApp.RtcSession.DataChannel && ViewerApp.RtcSession.DataChannel.readyState == "open") {
             rtcSend();
         }
-        else if (ViewerApp.ViewerHubConnection.Connection?.connectionStarted) {
+        else if (ViewerApp.ViewerHubConnection.Connection) {
             websocketSend();
+        }
+        else {
+            console.warn("No connection available.");
         }
     }
 
