@@ -1,7 +1,7 @@
 ﻿export type HubConnection = {
     start: () => Promise<any>;
+    onclose: (callback: () => any) => any;
     connectionStarted: boolean;
-    closedCallbacks: any[];
     invoke: (...rest) => any;
     stop: () => any;
 }
