@@ -23,7 +23,7 @@ export class ViewerHubConnection {
             UI.StatusMessage.innerHTML = `Connection error: ${err.message}`;
             UI.ToggleConnectUI(true);
         });
-        this.Connection.closedCallbacks.push((ev) => {
+        this.Connection.onclose(() => {
             UI.ToggleConnectUI(true);
         });
         ViewerApp.ClipboardWatcher.WatchClipboard();
